Packet Updating
Added and Updated packets for version 1.04.16

+++ SourceFiles/L2RPacketReader/L2RPacketReader/L2RPacketReader/CSV/class.cs
@@ -20,7 +20,7 @@ namespace L2RPacketReader.CSV
     class Class
     {
         private static IEnumerable<classheader> _records;
-        public static string className(UInt16 intID)
+        public static string className(UInt32 intID)
         {
             string result = string.Empty;
             string id = intID.ToString();
